Move impl_array_newtype_stringify to include
The impl_array_newtype_stringify macro is an internal macro in bitcoin
that is used in bip32 and other logic. As part of the separation of
bip32 functionality into a new crate, this macro will be needed.

Split impl_array_newtype_stringify into a module in include/ directory
and include! in existing location in bitcoin.
